requirements.txt
Django==1.7.6
django-taggit==0.12.3
djangorestframework==3.1.0
djangorestframework-jwt==1.3.2
mysqlclient==1.3.6
PyJWT==0.4.3
